refactoring de la fonction yatzy

+++ src/test/java/YatzyTest.java
@@ -13,10 +13,10 @@ public class YatzyTest {
 
     @Test public void yatzy_scores_50() {
         int expected = 50;
-        int actual = Yatzy.yatzy(4,4,4,4,4);
+        int actual = Yatzy.yatzy((byte)4, (byte)4, (byte)4, (byte)4, (byte)4);
         assertEquals(expected, actual);
-        assertEquals(50, Yatzy.yatzy(6,6,6,6,6));
-        assertEquals(0, Yatzy.yatzy(6,6,6,6,3));
+        assertEquals(50, Yatzy.yatzy((byte)6, (byte)6, (byte)6, (byte)6, (byte)6));
+        assertEquals(0, Yatzy.yatzy((byte)6, (byte)6, (byte)6, (byte)6, (byte)3));
     }
 
     @Test public void test_1s() {
